Refactored packages in pkg
