+++ pkg/storage/file.go
@@ -1,4 +1,4 @@
-package server
+package storage
 
 import (
 	"sync"
@@ -10,17 +10,6 @@ import (
     "gopkg.in/yaml.v2"
     "fmt"
 )
-
-type Storage interface {
-    ClaimTestcase(tester string, caseID string, claim bool) error
-    SetTestcaseRun(tester string, caseID string, result protocol.TestCaseResult, comment string) error
-    AddParticipant(tester string) error
-    Save() error
-
-    GetSuite() protocol.TestSuite
-    GetRun() protocol.TestRun
-    GetParticipant(tester string) (protocol.TestParticipant, bool)
-}
 
 type OnSaveFunc func(st *FileStorage)
